requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
amqp==5.3.1
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
celery==5.5.1
celery-healthcheck==0.1.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloud-sql-python-connector==1.18.1
cryptography==44.0.2
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.2.0
Faker==37.1.0
fastapi==0.115.12
flasgger==0.9.7b2
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
flask-openapi3==4.1.0
flask-openapi3-elements==9.0.1
flask-openapi3-rapidoc==9.3.8
flask-openapi3-rapipdf==2.2.1
flask-openapi3-redoc==2.5.0
flask-openapi3-scalar==1.28.19
flask-openapi3-swagger==5.21.0
Flask-RESTful==0.3.10
Flask-Session==0.8.0
flask-shell-ipython==0.5.3
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flower==2.0.1
frozenlist==1.6.0
google-auth==2.39.0
greenlet==3.2.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
humanize==4.12.2
idna==3.10
iniconfig==2.1.0
ipdb==0.13.13
ipython==8.35.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.3
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
msgspec==0.19.0
multidict==6.4.3
mysql-connector-python==9.2.0
openai==1.75.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pika==1.3.2
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.3.5
pytest-flask-sqlalchemy==1.1.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
yarl==1.20.0
